Move parser and ping to utils folder
